Improving test by mocking backend stub.
